requirements.txt
Django>=5.0,<6.0
django-admin-sortable2>=2.3.0
Pillow>=10.0
psycopg[binary]>=3.1
gunicorn>=21.0
python-dotenv>=1.0
openai>=1.0
deepl>=1.20.0
polib>=1.2.0
Babel>=2.14.0
